refactor: rename Files.php to File.php
